+++ src/renderer/components/wallet/txs/interact/Interact.types.ts
@@ -1,0 +1,1 @@
+export type InteractType = 'bond' | 'unbond' | 'leave' | 'custom'
