--- src/renderer/components/wallet/txs/interact/forms/Custom.tsx
@@ -1,73 +1,0 @@
-import React, { useCallback } from 'react'
-
-import { assetAmount, assetToBase, BaseAmount } from '@xchainjs/xchain-util'
-import { Form } from 'antd'
-import BigNumber from 'bignumber.js'
-import { useIntl } from 'react-intl'
-
-import { THORCHAIN_DECIMAL } from '../../../../../helpers/assetHelper'
-import { Input, InputBigNumber } from '../../../../uielements/input'
-import * as Styled from './Forms.styles'
-
-type FormValues = { memo: string; amount: BigNumber }
-
-type Props = {
-  onFinish: (boundData: { memo: string; amount: BaseAmount }) => void
-  isLoading?: boolean
-  loadingProgress?: string
-}
-export const Custom: React.FC<Props> = ({ onFinish: onFinishProp, isLoading = false, loadingProgress }) => {
-  const intl = useIntl()
-  const [form] = Form.useForm<FormValues>()
-
-  const onFinish = useCallback(
-    ({ memo, amount }: FormValues) => {
-      onFinishProp({
-        memo,
-        amount: assetToBase(assetAmount(amount))
-      })
-    },
-    [onFinishProp]
-  )
-
-  return (
-    <Styled.Form form={form} onFinish={onFinish} initialValues={{ memo: '' }}>
-      <div>
-        <Styled.InputContainer>
-          <Styled.InputLabel>{intl.formatMessage({ id: 'common.memo' })}</Styled.InputLabel>
-          <Form.Item
-            name="memo"
-            rules={[
-              {
-                required: true,
-                message: intl.formatMessage({ id: 'wallet.validations.shouldNotBeEmpty' })
-              }
-            ]}>
-            <Input disabled={isLoading} size="large" />
-          </Form.Item>
-        </Styled.InputContainer>
-
-        <Styled.InputContainer>
-          <Styled.InputLabel>{intl.formatMessage({ id: 'common.amount' })}</Styled.InputLabel>
-          <Form.Item name="amount">
-            <InputBigNumber disabled={isLoading} size="large" decimal={THORCHAIN_DECIMAL} />
-          </Form.Item>
-        </Styled.InputContainer>
-      </div>
-
-      <Styled.SubmitButtonContainer>
-        {() => (
-          <>
-            {loadingProgress && <Styled.LoadingProgress>{loadingProgress}</Styled.LoadingProgress>}
-            <Styled.SubmitButton
-              loading={isLoading}
-              disabled={isLoading || !!form.getFieldsError().filter(({ errors }) => errors.length).length}
-              htmlType="submit">
-              {intl.formatMessage({ id: 'wallet.action.send' })}
-            </Styled.SubmitButton>
-          </>
-        )}
-      </Styled.SubmitButtonContainer>
-    </Styled.Form>
-  )
-}
