--- src/renderer/components/wallet/txs/interact/forms/index.ts
@@ -1,1 +1,0 @@
-export * from './Bond'
