+++ src/renderer/i18n/types.ts
@@ -311,6 +311,7 @@ type DepositMessageKey =
   | 'deposit.interact.title'
   | 'deposit.interact.subtitle'
   | 'deposit.interact.actions'
+  // Note: Never change last keys (bond, unbond, leave, custom) - they are mapped to `InteractTypes`
   | 'deposit.interact.actions.bond'
   | 'deposit.interact.actions.unbond'
   | 'deposit.interact.actions.leave'
