--- src/renderer/views/wallet/Interact/BondView.tsx
@@ -1,73 +1,0 @@
-import React, { useMemo } from 'react'
-
-import { THORChain } from '@xchainjs/xchain-util'
-import * as FP from 'fp-ts/function'
-import * as O from 'fp-ts/Option'
-import { useObservableState } from 'observable-hooks'
-
-import { WalletType } from '../../../../shared/wallet/types'
-import { LoadingView } from '../../../components/shared/loading'
-import { Bond } from '../../../components/wallet/txs/interact/forms'
-import { useThorchainContext } from '../../../contexts/ThorchainContext'
-import { useWalletContext } from '../../../contexts/WalletContext'
-import { getWalletBalanceByAddress } from '../../../helpers/walletHelper'
-import { useNetwork } from '../../../hooks/useNetwork'
-import { useOpenExplorerTxUrl } from '../../../hooks/useOpenExplorerTxUrl'
-import { useValidateAddress } from '../../../hooks/useValidateAddress'
-import { DEFAULT_BALANCES_FILTER, INITIAL_BALANCES_STATE } from '../../../services/wallet/const'
-
-type Props = {
-  walletType: WalletType
-  walletIndex: number
-  walletAddress: string
-}
-
-export const BondView: React.FC<Props> = (props) => {
-  const { walletType, walletIndex, walletAddress } = props
-
-  const { network } = useNetwork()
-  const {
-    balancesState$,
-    keystoreService: { validatePassword$ }
-  } = useWalletContext()
-
-  const [{ balances: oBalances }] = useObservableState(
-    () => balancesState$(DEFAULT_BALANCES_FILTER),
-    INITIAL_BALANCES_STATE
-  )
-
-  const { openExplorerTxUrl, getExplorerTxUrl } = useOpenExplorerTxUrl(O.some(THORChain))
-
-  const { validateAddress } = useValidateAddress(THORChain)
-
-  const oWalletBalance = useMemo(
-    () =>
-      FP.pipe(
-        oBalances,
-        O.chain((balances) => getWalletBalanceByAddress(balances, walletAddress))
-      ),
-    [oBalances, walletAddress]
-  )
-
-  const { interact$ } = useThorchainContext()
-
-  return FP.pipe(
-    oWalletBalance,
-    O.fold(
-      () => <LoadingView size="large" />,
-      (walletBalance) => (
-        <Bond
-          walletIndex={walletIndex}
-          walletType={walletType}
-          balance={walletBalance}
-          interact$={interact$}
-          openExplorerTxUrl={openExplorerTxUrl}
-          getExplorerTxUrl={getExplorerTxUrl}
-          addressValidation={validateAddress}
-          validatePassword$={validatePassword$}
-          network={network}
-        />
-      )
-    )
-  )
-}
